+++ app/(datasets)/datasets/page.tsx
@@ -2,32 +2,11 @@ import React from 'react';
 import { getDatasets } from 'app/blog/utils/mdx';
 import { Suspense } from 'react';
 import Catalog from './catalog';
+import { transformData } from '@helpers/data';
 
 export default function Page() {
-  const posts: any[] = getDatasets();
-
-  const transformData = () => {
-    const data = posts?.map((post) => ({
-      ...post.metadata,
-    }));
-
-    const result = data?.map((d) => {
-      const updatedTax = d.taxonomy.map((t) => {
-        const updatedVals = t.values.map((v) => {
-          return {
-            id: v.replace(/ /g, '_').toLowerCase(),
-            name: v,
-          };
-        });
-        return { ...t, values: updatedVals };
-      });
-      return { ...d, taxonomy: updatedTax };
-    });
-
-    return result;
-  };
-
-  const transformed = transformData();
+  const posts: any[] = getDatasets(); // @TODO: Revist type here, should use data types from veda-ui
+  const transformed = transformData(posts);
 
   return (
     <section>
