+++ app/store/providers/theme.tsx
@@ -1,7 +1,7 @@
 'use client';
 import React, { ReactNode } from 'react';
 import { createUITheme } from '@devseed-ui/theme-provider';
-import { DevseedUiThemeProvider, PageMainContent } from '../../lib';
+import { DevseedUiThemeProvider, PageMainContent } from '@lib';
 
 const VEDA_OVERRIDE_THEME = {
   zIndices: {
