+++ tsconfig.json
@@ -15,6 +15,11 @@
     "jsx": "preserve",
     "baseUrl": ".",
     "incremental": true,
+    "paths": {
+      "@lib": ["app/lib"],
+      "@lib/*": ["app/lib/*"],
+      "@helpers/*": ["app/utilities/helpers/*"]
+    },
     "plugins": [
       {
         "name": "next"
